requirements.txt
python-dotenv==0.17.*
requests==2.25.*